+++ source/content/item/weapon/cannon/BoulderBomber.java
@@ -15,7 +15,7 @@ import javax.annotation.Nullable;
 import java.util.List;
 
 public class BoulderBomber extends BaseCannon {
-	public BoulderBomber(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public BoulderBomber(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
