+++ source/content/item/weapon/cannon/CarrotCannon.java
@@ -13,7 +13,7 @@ import net.tslat.aoa3.content.entity.projectile.gun.BaseBullet;
 import javax.annotation.Nullable;
 
 public class CarrotCannon extends BaseCannon {
-	public CarrotCannon(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public CarrotCannon(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
