+++ source/content/item/weapon/cannon/ClownCannon.java
@@ -22,7 +22,7 @@ import javax.annotation.Nullable;
 import java.util.List;
 
 public class ClownCannon extends BaseCannon {
-	public ClownCannon(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public ClownCannon(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
