+++ source/content/item/weapon/cannon/CoralCannon.java
@@ -19,7 +19,7 @@ import javax.annotation.Nullable;
 import java.util.List;
 
 public class CoralCannon extends BaseCannon {
-	public CoralCannon(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public CoralCannon(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
