+++ source/content/item/weapon/cannon/GolderBomber.java
@@ -16,7 +16,7 @@ import javax.annotation.Nullable;
 import java.util.List;
 
 public class GolderBomber extends BaseCannon {
-	public GolderBomber(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public GolderBomber(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
