+++ source/content/item/weapon/cannon/IroCannon.java
@@ -16,7 +16,7 @@ import javax.annotation.Nullable;
 import java.util.List;
 
 public class IroCannon extends BaseCannon {
-	public IroCannon(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public IroCannon(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
