+++ source/content/item/weapon/cannon/JackFunger.java
@@ -16,7 +16,7 @@ public class JackFunger extends BaseCannon {
 	double dmg;
 	int firingDelay;
 
-	public JackFunger(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public JackFunger(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 		this.dmg = dmg;
 		this.firingDelay = firingDelayTicks;
