+++ source/content/item/weapon/cannon/MechaCannon.java
@@ -11,7 +11,7 @@ import net.tslat.aoa3.content.entity.projectile.gun.BaseBullet;
 import javax.annotation.Nullable;
 
 public class MechaCannon extends BaseCannon {
-	public MechaCannon(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public MechaCannon(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
