+++ source/content/item/weapon/cannon/PredatorianBlaster.java
@@ -21,7 +21,7 @@ import javax.annotation.Nullable;
 import java.util.List;
 
 public class PredatorianBlaster extends BaseCannon {
-	public PredatorianBlaster(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public PredatorianBlaster(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
