+++ source/content/item/weapon/cannon/ShadowBlaster.java
@@ -11,7 +11,7 @@ import net.tslat.aoa3.content.entity.projectile.gun.BaseBullet;
 import javax.annotation.Nullable;
 
 public class ShadowBlaster extends BaseCannon {
-	public ShadowBlaster(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public ShadowBlaster(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
