+++ source/content/item/weapon/cannon/UltraCannon.java
@@ -11,7 +11,7 @@ import net.tslat.aoa3.content.entity.projectile.gun.BaseBullet;
 import javax.annotation.Nullable;
 
 public class UltraCannon extends BaseCannon {
-	public UltraCannon(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public UltraCannon(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
