+++ source/content/item/weapon/cannon/VoxCannon.java
@@ -24,7 +24,7 @@ public class VoxCannon extends BaseCannon {
 	double dmg;
 	int firingDelay;
 
-	public VoxCannon(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public VoxCannon(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 		this.dmg = dmg;
 		this.firingDelay = firingDelayTicks;
