+++ source/content/item/weapon/cannon/WitherCannon.java
@@ -21,7 +21,7 @@ import javax.annotation.Nullable;
 import java.util.List;
 
 public class WitherCannon extends BaseCannon {
-	public WitherCannon(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public WitherCannon(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
