+++ source/content/item/weapon/gun/Artifact.java
@@ -13,7 +13,7 @@ import javax.annotation.Nullable;
 import java.util.List;
 
 public class Artifact extends BaseGun {
-	public Artifact(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public Artifact(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
