+++ source/content/item/weapon/gun/ChilliChugger.java
@@ -23,7 +23,7 @@ public class ChilliChugger extends BaseGun {
 	double dmg;
 	int firingDelay;
 
-	public ChilliChugger(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public ChilliChugger(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 		this.dmg = dmg;
 		this.firingDelay = firingDelayTicks;
