+++ source/content/item/weapon/gun/Clownershot.java
@@ -15,7 +15,7 @@ import javax.annotation.Nullable;
 import java.util.List;
 
 public class Clownershot extends BaseGun {
-	public Clownershot(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public Clownershot(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
