+++ source/content/item/weapon/gun/Construct.java
@@ -6,7 +6,7 @@ import net.tslat.aoa3.common.registration.AoASounds;
 import javax.annotation.Nullable;
 
 public class Construct extends BaseGun {
-	public Construct(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public Construct(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
