+++ source/content/item/weapon/gun/CoralClogger.java
@@ -24,7 +24,7 @@ import javax.annotation.Nullable;
 import java.util.List;
 
 public class CoralClogger extends BaseGun {
-	public CoralClogger(double dmg, int durability, int firingDelayTicks, float recoil) {
+	public CoralClogger(float dmg, int durability, int firingDelayTicks, float recoil) {
 		super(dmg, durability, firingDelayTicks, recoil);
 	}
